initial cursor store (#2482)
based off the first implementation #2262

cursor store/vector clock traits

d14n needs to know last seen of all originators/topics for a query to be correct.

+++ xmtp_configuration/src/common/d14n.rs
@@ -2,12 +2,12 @@
 pub struct Originators;
 
 impl Originators {
-    pub const MLS_COMMITS: u16 = 0;
-    pub const INBOX_LOG: u16 = 1;
+    pub const MLS_COMMITS: u32 = 0;
+    pub const INBOX_LOG: u32 = 1;
 
-    pub const APPLICATION_MESSAGES: u16 = 10;
-    pub const WELCOME_MESSAGES: u16 = 11;
+    pub const APPLICATION_MESSAGES: u32 = 10;
+    pub const WELCOME_MESSAGES: u32 = 11;
     /// Key Packages
-    pub const INSTALLATIONS: u16 = 13;
-    pub const REMOTE_COMMIT_LOG: u16 = 100;
+    pub const INSTALLATIONS: u32 = 13;
+    pub const REMOTE_COMMIT_LOG: u32 = 100;
 }
